+++ src/main.ts
@@ -8,6 +8,7 @@ import {
   Plugin,
   PluginSettingTab,
   Setting,
+  stringifyYaml,
   TFile,
   TFolder,
 } from "obsidian";
@@ -15,11 +16,12 @@ import {
   Article,
   compareHighlightsInFile,
   DATE_FORMAT,
+  formatDate,
   getHighlightLocation,
   loadArticles,
   PageType,
   parseDateTime,
-  unicodeSlug,
+  replaceIllegalChars,
 } from "./util";
 import { FolderSuggest } from "./settings/file-suggest";
 
@@ -57,28 +59,11 @@ const DEFAULT_SETTINGS: Settings = {
   filter: "HIGHLIGHTS",
   syncAt: "",
   customQuery: "",
-  template: `---
-title: {{{title}}}
-{{#author}}
-author: {{{author}}}
-{{/author}}
-{{#labels.length}}
-tags:
-{{#labels}}  - {{{name}}}
-{{/labels}}
-{{/labels.length}}
-date_saved: {{{dateSaved}}}
----
-
-# {{{title}}}
+  template: `# {{{title}}}
 #Omnivore
 
 [Read on Omnivore]({{{omnivoreUrl}}})
 [Read Original]({{{originalUrl}}})
-{{#content}}
-
-{{{content}}}
-{{/content}}
 
 {{#highlights.length}}
 ## Highlights
@@ -181,7 +166,7 @@ export default class OmnivorePlugin extends Plugin {
     await this.saveSettings();
 
     try {
-      console.log(`obsidian-omnivore starting sync since: '${syncAt}`);
+      console.log(`obsidian-omnivore starting sync since: '${syncAt}'`);
 
       new Notice("🚀 Fetching articles ...");
 
@@ -203,12 +188,8 @@ export default class OmnivorePlugin extends Plugin {
         );
 
         for (const article of articles) {
-          const dateSaved = DateTime.fromISO(article.savedAt).toFormat(
-            this.settings.dateSavedFormat
-          );
-          const subFolderName = DateTime.fromISO(article.savedAt).toFormat(
-            this.settings.folderFormat
-          );
+          const dateFormat = this.settings.dateSavedFormat;
+          const subFolderName = formatDate(article.savedAt, this.settings.folderFormat);
           let folderName;
           if (this.settings.useFolder) {
             folderName = `${folder}/${subFolderName}`;
@@ -221,15 +202,6 @@ export default class OmnivorePlugin extends Plugin {
           if (!(omnivoreFolder instanceof TFolder)) {
             await this.app.vault.createFolder(folderName);
           }
-
-          // use unicode slug to show characters from other languages in the file name
-          const pageName = `${folderName}/${unicodeSlug(
-            article.title,
-            article.savedAt
-          )}.md`;
-          const siteName =
-            article.siteName ||
-            this.siteNameFromUrl(article.originalArticleUrl);
 
           // sort highlights by location if selected in options
           highlightOrder === "LOCATION" &&
@@ -248,7 +220,6 @@ export default class OmnivorePlugin extends Plugin {
                 return compareHighlightsInFile(a, b);
               }
             });
-
           const highlights = article.highlights?.map((highlight) => {
             return {
               text: highlight.quote,
@@ -259,9 +230,13 @@ export default class OmnivorePlugin extends Plugin {
               note: highlight.annotation,
             };
           });
-
+          const dateSaved = formatDate(article.savedAt, dateFormat);
+          const siteName =
+            article.siteName ||
+            this.siteNameFromUrl(article.originalArticleUrl);
           // Build content string based on template
           const content = Mustache.render(template, {
+            id: article.id,
             title: article.title,
             omnivoreUrl: `https://omnivore.app/me/${article.slug}`,
             siteName,
@@ -276,16 +251,91 @@ export default class OmnivorePlugin extends Plugin {
             highlights,
             content: article.content,
           });
-
+          const publishedAt = article.publishedAt;
+          const datePublished = publishedAt
+            ? formatDate(publishedAt, dateFormat)
+            : null;
+          // add frontmatter to the content
+          const frontmatter = {
+            id: article.id,
+            title: article.title,
+            author: article.author,
+            tags: article.labels?.map((l) => l.name),
+            date_saved: dateSaved,
+            date_published: datePublished,
+          };
+          // remove null and empty values from frontmatter
+          const filteredFrontmatter = Object.fromEntries(
+            Object.entries(frontmatter).filter(
+              ([_, value]) => value != null && value !== ""
+            )
+          );
+          const frontmatterYaml = stringifyYaml(filteredFrontmatter);
+          const frontmatterString = `---\n${frontmatterYaml}---`;
+          // Modify the contents of the note with the updated front matter
+          let updatedContent = content.replace(
+            /^(---[\s\S]*?---)/gm,
+            frontmatterString
+          );
+          // if the content doesn't have frontmatter, add it
+          if (!content.match(/^(---[\s\S]*?---)/gm)) {
+            updatedContent = `${frontmatterString}\n\n${content}`;
+          }
+          // use the title as the filename
+          const pageName = `${folderName}/${replaceIllegalChars(
+            article.title
+          )}.md`;
           const normalizedPath = normalizePath(pageName);
           const omnivoreFile = app.vault.getAbstractFileByPath(normalizedPath);
-          if (omnivoreFile instanceof TFile) {
-            const existingContent = await this.app.vault.read(omnivoreFile);
-            if (existingContent !== content) {
-              await this.app.vault.modify(omnivoreFile, content);
+          try {
+            if (omnivoreFile instanceof TFile) {
+              await app.fileManager.processFrontMatter(
+                omnivoreFile,
+                async (frontMatter) => {
+                  const id = frontMatter.id;
+                  if (id && id !== article.id) {
+                    // this article has the same name but different id
+                    const newPageName = `${folderName}/${replaceIllegalChars(
+                      article.title
+                    )}-${article.id}.md`;
+                    const newNormalizedPath = normalizePath(newPageName);
+                    const newOmnivoreFile =
+                      app.vault.getAbstractFileByPath(newNormalizedPath);
+                    if (newOmnivoreFile instanceof TFile) {
+                      // a file with the same name and id already exists, so we need to update it
+                      const existingContent = await this.app.vault.read(
+                        newOmnivoreFile
+                      );
+                      if (existingContent !== updatedContent) {
+                        await this.app.vault.modify(
+                          newOmnivoreFile,
+                          updatedContent
+                        );
+                      }
+                      return;
+                    }
+                    // a file with the same name but different id already exists, so we need to create it
+                    await this.app.vault.create(
+                      newNormalizedPath,
+                      updatedContent
+                    );
+                    return;
+                  }
+                  // a file with the same id already exists, so we might need to update it
+                  const existingContent = await this.app.vault.read(
+                    omnivoreFile
+                  );
+                  if (existingContent !== updatedContent) {
+                    await this.app.vault.modify(omnivoreFile, updatedContent);
+                  }
+                }
+              );
+            } else if (!omnivoreFile) {
+              // file doesn't exist, so we need to create it
+              await this.app.vault.create(normalizedPath, updatedContent);
             }
-          } else if (!omnivoreFile) {
-            await this.app.vault.create(normalizedPath, content);
+          } catch (e) {
+            console.error(e);
           }
         }
       }
@@ -446,7 +496,10 @@ class OmnivoreSettingTab extends PluginSettingTab {
             fragment.createEl("a", {
               text: "Reference",
               href: "https://github.com/janl/mustache.js/#templates",
-            })
+            }),
+            fragment.createEl("br"),
+            fragment.createEl("br"),
+            "Available variables: id, title, omnivoreUrl, siteName, originalUrl, author, content, dateSaved, labels.name, highlights.text, highlights.highlightUrl, highlights.note, highlights.dateHighlighted"
           );
         })
       )
@@ -455,7 +508,11 @@ class OmnivoreSettingTab extends PluginSettingTab {
           .setPlaceholder("Enter the template")
           .setValue(this.plugin.settings.template)
           .onChange(async (value) => {
-            this.plugin.settings.template = value;
+            // TODO: validate template
+            // if template is empty, use default template
+            this.plugin.settings.template = value
+              ? value
+              : DEFAULT_SETTINGS.template;
             await this.plugin.saveSettings();
           })
         text.inputEl.setAttr("rows", 10);